Move pairlists to be a plugin submodule
