+++ freqtrade/plugins/pairlist/IPairList.py
