+++ freqtrade/plugins/pairlist/SpreadFilter.py
@@ -4,7 +4,7 @@ Spread pair list filter
 import logging
 from typing import Any, Dict
 
-from freqtrade.pairlist.IPairList import IPairList
+from freqtrade.plugins.pairlist.IPairList import IPairList
 
 
 logger = logging.getLogger(__name__)
