+++ freqtrade/plugins/pairlist/StaticPairList.py
@@ -7,7 +7,7 @@ import logging
 from typing import Any, Dict, List
 
 from freqtrade.exceptions import OperationalException
-from freqtrade.pairlist.IPairList import IPairList
+from freqtrade.plugins.pairlist.IPairList import IPairList
 
 
 logger = logging.getLogger(__name__)
