+++ freqtrade/plugins/pairlist/VolumePairList.py
@@ -8,7 +8,7 @@ from datetime import datetime
 from typing import Any, Dict, List
 
 from freqtrade.exceptions import OperationalException
-from freqtrade.pairlist.IPairList import IPairList
+from freqtrade.plugins.pairlist.IPairList import IPairList
 
 
 logger = logging.getLogger(__name__)
