+++ freqtrade/plugins/pairlistmanager.py
@@ -9,7 +9,7 @@ from cachetools import TTLCache, cached
 
 from freqtrade.constants import ListPairsWithTimeframes
 from freqtrade.exceptions import OperationalException
-from freqtrade.pairlist.IPairList import IPairList
+from freqtrade.plugins.pairlist.IPairList import IPairList
 from freqtrade.resolvers import PairListResolver
